requirements.txt
boto3==1.16.22
django-heroku==0.3.1
django-storages==1.10.1
gunicorn==20.0.4
ipython==7.19.0
psycopg2==2.8.6
python-dotenv==0.15.0
